requirements.txt

viam-sdk==0.50.0
typing-extensions
numpy>=1.26.4
torchvision
torch
pillow
protobuf>=4.25.1
pyinstaller


